requirements.txt
torch
torchvision
torchmetrics
numpy
pandas
scikit-learn
matplotlib
seaborn
tqdm
Pillow
opencv-python
fastapi
uvicorn[standard]
python-multipart
